+++ src/components/button.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Slot } from '@radix-ui/react-slot'
 
-import { cn } from 'utils'
+import { cn } from 'utils/cn'
 
 import { cva, type VariantProps } from 'class-variance-authority'
