+++ src/components/slider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import * as SliderPrimitive from '@radix-ui/react-slider'
 
-import { cn } from 'utils'
+import { cn } from 'utils/cn'
 
 const Slider = React.forwardRef<
   React.ElementRef<typeof SliderPrimitive.Root>,
